make main content span full width of screen when sidebar is closed
Allmost completed making maincontetn span the width of the screen when sidebar is closed. I first has to fix the hamburge. The Burger was located in the sidebar and postitoned into the header. I moved the burger into the header. I then set props and created functions so that the open and closing of sidebar still worked. Then i started working on the maincontetn css. left off with probelm of an infinate loop.