remove chatRoom function
Added the ability to remove a chatRoom to messenger component. i have isolated test the function and also tested it while passing hardcoded values to it. Once sidebar is fixed we can call the function from sidebar on click and verify it works with more dynamic data. Also did a couple comments to authentication