auth compnent redux store
subscribed teh auth component to the redux store. Created a action and reducer to set authentication in store and got setting the auth working far as i know. also added logger and logger is logging when the auth state changes. The auth component is not running off of redux yet.