redux state authentcation component
authenticaiont component now uses redux state. In authentication component there is still state for the notificiation as i am not sure if we want to put that in redux state yet. Currently you can not log into the app as when it gets to messenger messenger redirects you back to auth. something to do with messenger not being converted to redux yet. app.js now also uses redux. I had to setup the component to use the state. it was allready subscried to the store but was not using it.