Account settings username and password update form styles.
Changed the styles for the update username and update password form in account settings. I reduced the padding to allow the form to display better on smaller desktop screens. I also removed an unecessary div inside of UpdatUsernameForm that was causing a scroll bar when their was nothing to scroll to.